@@ -1,6 +1,7 @@
 package cucumberexpressions
 
 import (
+	"errors"
 	"fmt"
 	"reflect"
 	"regexp"
@@ -22,6 +23,10 @@ type CucumberExpression struct {
 
 func NewCucumberExpression(expression string, parameterTypeRegistry *ParameterTypeRegistry) (*CucumberExpression, error) {
 	result := &CucumberExpression{source: expression, parameterTypeRegistry: parameterTypeRegistry}
+
+	if strings.Contains(expression, `(?!`) {
+		return nil, errors.New("sorry, go does not support ?! used as a regex negative matcher")
+	}
 
 	expression = result.processEscapes(expression)
 
@@ -114,9 +119,9 @@ func (c *CucumberExpression) processAlteration(expression string) (string, error
 				}
 			}
 			return fmt.Sprintf("(?:%s)", replacement)
-		} else {
-			return replacement
 		}
+
+		return replacement
 	})
 	return result, err
 }
