@@ -6,24 +6,24 @@ import java.util.Iterator;
 
 /**
  * Iterates over messages read from a stream. Client code should not depend on this class
- * directly, but rather on a {@code Iterable<Messages.Wrapper>} object.
- * Tests can then use a {@code new ArrayList<Messages.Wrapper>} which implements the same interface .
+ * directly, but rather on a {@code Iterable<Messages.Envelope>} object.
+ * Tests can then use a {@code new ArrayList<Messages.Envelope>} which implements the same interface .
  */
-public class ProtobufStreamIterable implements Iterable<Messages.Wrapper> {
+public class ProtobufStreamIterable implements Iterable<Messages.Envelope> {
     private final InputStream input;
-    private Messages.Wrapper next;
+    private Messages.Envelope next;
 
     public ProtobufStreamIterable(InputStream input) {
         this.input = input;
     }
 
     @Override
-    public Iterator<Messages.Wrapper> iterator() {
-        return new Iterator<Messages.Wrapper>() {
+    public Iterator<Messages.Envelope> iterator() {
+        return new Iterator<Messages.Envelope>() {
             @Override
             public boolean hasNext() {
                 try {
-                    next = Messages.Wrapper.parseDelimitedFrom(input);
+                    next = Messages.Envelope.parseDelimitedFrom(input);
                     return next != null;
                 } catch (IOException e) {
                     throw new RuntimeException(e);
@@ -31,7 +31,7 @@ public class ProtobufStreamIterable implements Iterable<Messages.Wrapper> {
             }
 
             @Override
-            public Messages.Wrapper next() {
+            public Messages.Envelope next() {
                 if(next == null) {
                     throw new IllegalStateException("next() should only be called after a call to hasNext() that returns true");
                 }
