@@ -6,7 +6,7 @@ module Cucumber
       def self.call(io)
         Enumerator.new do |yielder|
           while !io.eof?
-            yielder.yield(Cucumber::Messages::Wrapper.parse_delimited_from(io))
+            yielder.yield(Cucumber::Messages::Envelope.parse_delimited_from(io))
           end
         end
       end
