@@ -38,7 +38,7 @@ module Gherkin
       stdin, stdout, stderr, wait_thr = Open3.popen3(*args)
       stdin.binmode
       @sources.each do |source|
-        wrapper = Cucumber::Messages::Wrapper.new(
+        wrapper = Cucumber::Messages::Envelope.new(
           source: source
         )
         wrapper.write_delimited_to(stdin)
@@ -60,7 +60,7 @@ module Gherkin
 
     def self.encode_source_message(uri, data)
       media_obj = Cucumber::Messages::Media.new
-      media_obj.encoding = 'UTF-8'
+      media_obj.encoding = :UTF8
       media_obj.content_type = 'text/x.cucumber.gherkin+plain'
       source_obj = Cucumber::Messages::Source.new
       source_obj.uri = uri
